+++ src/assignment/mod.rs
@@ -9,24 +9,24 @@ mod group_management;
 ///
 /// Assignments are typically obtained from an [Assigner](assigners::Assigner)
 pub struct Assignment {
-    subject_ids_to_group_ids: HashMap<u64, u64>,
-    group_ids_to_subjects_ids: HashMap<u64, Vec<u64>>,
+    subject_ids_to_group_ids: HashMap<u32, u32>,
+    group_ids_to_subjects_ids: HashMap<u32, Vec<u32>>,
 }
 impl Assignment {
     /// Get the id of the group the given subject is assigned to.
-    pub fn subject_to_group_id<S: Subject>(&self, subject: &S) -> Option<&u64> {
+    pub fn subject_to_group_id<S: Subject>(&self, subject: &S) -> Option<&u32> {
         self.subject_ids_to_group_ids.get(&subject.id())
     }
     /// Get the ids of the subjects assigned to the given group.
-    pub fn group_to_subjects_ids<G: Group>(&self, group: &G) -> Option<&Vec<u64>> {
+    pub fn group_to_subjects_ids<G: Group>(&self, group: &G) -> Option<&Vec<u32>> {
         self.group_ids_to_subjects_ids.get(&group.id())
     }
 }
 
 impl Default for Assignment {
     fn default() -> Self {
-        let subject_ids_to_group_ids: HashMap<u64, u64> = HashMap::new();
-        let group_ids_to_subjects_ids: HashMap<u64, Vec<u64>> = HashMap::new();
+        let subject_ids_to_group_ids: HashMap<u32, u32> = HashMap::new();
+        let group_ids_to_subjects_ids: HashMap<u32, Vec<u32>> = HashMap::new();
         Self {
             subject_ids_to_group_ids,
             group_ids_to_subjects_ids,
@@ -34,8 +34,8 @@ impl Default for Assignment {
     }
 }
 
-impl From<(HashMap<u64, u64>, HashMap<u64, Vec<u64>>)> for Assignment {
-    fn from(pair_of_maps: (HashMap<u64, u64>, HashMap<u64, Vec<u64>>)) -> Self {
+impl From<(HashMap<u32, u32>, HashMap<u32, Vec<u32>>)> for Assignment {
+    fn from(pair_of_maps: (HashMap<u32, u32>, HashMap<u32, Vec<u32>>)) -> Self {
         Self {
             subject_ids_to_group_ids: pair_of_maps.0,
             group_ids_to_subjects_ids: pair_of_maps.1,
@@ -45,8 +45,8 @@ impl From<(HashMap<u64, u64>, HashMap<u64, Vec<u64>>)> for Assignment {
 
 impl From<Assignment>
     for (
-        std::collections::HashMap<u64, u64>,
-        std::collections::HashMap<u64, Vec<u64>>,
+        std::collections::HashMap<u32, u32>,
+        std::collections::HashMap<u32, Vec<u32>>,
     )
 {
     fn from(assignment: Assignment) -> Self {
