+++ src/subjects/test_utils.rs
@@ -2,13 +2,13 @@
 use crate::subjects::Subject;
 #[derive(Debug)]
 pub(crate) struct TestSubject {
-    id: u64,
-    preferences: Vec<u64>,
-    assigned_group_id: Option<u64>,
+    id: u32,
+    preferences: Vec<u32>,
+    assigned_group_id: Option<u32>,
 }
 
 impl TestSubject {
-    pub fn new(id: u64, preferences: Vec<u64>) -> TestSubject {
+    pub fn new(id: u32, preferences: Vec<u32>) -> TestSubject {
         TestSubject {
             id,
             preferences,
@@ -18,12 +18,12 @@ impl TestSubject {
 }
 
 impl Subject for TestSubject {
-    fn dissatisfaction(&self, group_id: &u64) -> u32 {
+    fn dissatisfaction(&self, group_id: &u32) -> u32 {
         let dissatisfaction = self.preferences.iter().position(|x| x == group_id);
         dissatisfaction.unwrap_or(self.preferences.len()) as u32
     }
 
-    fn id(&self) -> u64 {
+    fn id(&self) -> u32 {
         self.id
     }
 }
